refac out help view logic
